# openapi.yaml
openapi: 3.1.0
info:
  title: Country City DB API
  description: A comprehensive API for querying countries, states, cities, regions, and subregions worldwide.
  version: 1.0.0
servers:
  - url: /
paths:
...
  /api/v1/regions/{id}/subregions:
    get:
      summary: List subregions in a region
      operationId: listRegionSubregions
      tags: [Regions]
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/search"
        - $ref: "#/components/parameters/name"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/no_page"
      responses:
        "200":
          description: Paginated list of subregions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedSubregions"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      summary: List subregions in a region (JSON body)
      operationId: listRegionSubregionsPost
      tags: [Regions]
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryParams"
      responses:
        "200":
          description: Paginated list of subregions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedSubregions"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/subregions:
    get:
      summary: List subregions
      operationId: listSubregions
      tags: [Subregions]
      parameters:
        - $ref: "#/components/parameters/search"
        - $ref: "#/components/parameters/name"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/no_page"
      responses:
        "200":
          description: Paginated list of subregions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedSubregions"
    post:
      summary: List subregions (JSON body)
      operationId: listSubregionsPost
      tags: [Subregions]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryParams"
      responses:
        "200":
          description: Paginated list of subregions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedSubregions"
  /api/v1/subregions/{id}:
    get:
      summary: Get subregion by ID
      operationId: getSubregion
      tags: [Subregions]
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Subregion object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subregion"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Countries ────────────────────────────────────────────
...
components:
  securitySchemes:
    internalKey:
      type: apiKey
      in: header
      name: X-Internal-Key
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Resource ID
...
    search:
      name: search
      in: query
      schema:
        type: string
      description: Full-text search query
    name:
      name: name
      in: query
      schema:
        type: string
      description: Filter by exact name
...
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
      description: Maximum number of results (max 100)
    offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
      description: Number of results to skip
    no_page:
      name: no_page
      in: query
      schema:
        type: boolean
        default: false
      description: If true, return all results without pagination
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    QueryParams:
      type: object
      properties:
        search:
          type: string
        name:
          type: string
        iso2:
          type: string
        iso3:
          type: string
        limit:
          type: integer
          default: 20
        offset:
          type: integer
          default: 0
        no_page:
          type: boolean
          default: false
...
    Subregion:
      type: object
      required: [id, name, region_id, updated_at, flag]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        translations:
          type: string
        region_id:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        flag:
          type: integer
        wiki_data_id:
          type: string
...
    PaginatedSubregions:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Subregion"
        total:
          type: integer
          format: int64
        limit:
          type: integer
        offset:
          type: integer
...